+++ src/pages/merge-summary/types.d.ts
@@ -1,0 +1,17 @@
+import EntityConfig from 'entities';
+
+export interface MergeProps {
+    sections: string;
+    newName: string;
+}
+
+interface EntityTableHeadProps<T> {
+    entity: EntityConfig<T>;
+}
+
+export interface EntityTableProps<T> {
+    entity: EntityConfig<T>;
+    list: T[];
+    selectedSections: number[];
+    newSectionName: string;
+}
